flask-server/requirements.txt
Flask
authlib
python-dotenv
